tacky: refactoring massive match expression for Expr into modules.
